// src/lib/pwa/store/pwa/types.ts
/**
 * The BeforeInstallPromptEvent is fired at the Window.onbeforeinstallprompt handler
 * before a user is prompted to "install" a website to a home screen on mobile.
 *
 */
export interface BeforeInstallPromptEvent extends Event {
  /**
   * Returns an array of DOMString items containing the platforms on which the event was dispatched.
   * This is provided for user agents that want to present a choice of versions to the user such as,
   * for example, "web" or "play" which would allow the user to chose between a web version or
   * an Android version.
   */
  readonly platforms: Array<string>

  /**
   * Returns a Promise that resolves to a DOMString containing either "accepted" or "dismissed".
   */
  readonly userChoice: Promise<{
    outcome: 'accepted' | 'dismissed'
    platform: string
  }>

  /**
   * Allows a developer to show the install prompt at a time of their own choosing.
   * This method returns a Promise.
   */
  prompt(): Promise<void>
}

export interface IState {
  loading: boolean
  installPromptEvent: BeforeInstallPromptEvent | undefined | null
  appInstalled: boolean
  choiceOutcome: 'accepted' | 'dismissed' | undefined
  standaloneQueryString: boolean
  standaloneMediaQuery: boolean
  fullscreenMediaQuery: boolean
  errorCode: string
  errorMsg: string
}

export interface Getters {}

export interface Actions {
  installApp(): Promise<void>
}